Update upstream libs to v2.1.0

+++ protobuf_srcs/tensorflow/lite/kernels/internal/types.h
@@ -1040,11 +1040,11 @@ struct SqueezeParams {
 
 struct StridedSliceParams {
   int8 start_indices_count;
-  int32 start_indices[4];
+  int16 start_indices[4];
   int8 stop_indices_count;
-  int32 stop_indices[4];
+  int16 stop_indices[4];
   int8 strides_count;
-  int32 strides[4];
+  int16 strides[4];
 
   int16 begin_mask;
   int16 ellipsis_mask;
